refactor(liferay-cli): move source code generation and build to a facet
