+++ projects/js-toolkit/packages/liferay-cli/src/new/facet-react/templates/.babelrc.ejs
